split up settings and profile into different pages. put in a burger menu and cleaned up the style a little
